# modules/services/ogc-api-records/openapi.yaml
openapi: 3.0.1
info:
  title: A sample API conforming to the OGC API - Records standard
  description: 'This is a sample OpenAPI definition that conforms to the OGC API - Records Service specification (conformance classes: "Core", "GeoJSON", "HTML" and "OpenAPI 3.0").'
  contact:
    name: Acme Corporation
    url: 'http://example.org/'
    email: info@example.org
  license:
    name: CC-BY 4.0 license
    url: 'https://creativecommons.org/licenses/by/4.0/'
  version: 1.0.0
servers:
  - url: 'https://virtserver.swaggerhub.com/constantinius/ogc-api-coverages/1.0.0'
    description: SwaggerHub API Auto Mocking
  - url: 'https://dev.example.org/'
    description: Development server
  - url: 'https://data.example.org/'
    description: Production server
...
paths:
...
  '/collections/{collectionId}':
    get:
      tags:
        - Collection
      summary: 'describe the {collectionId} collection'
      operationId: describeCollection
      parameters:
        - name: collectionId
          in: path
          description: Identifier (name) of a specific collection
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          description: 'Metadata about the {collectionId} collection shared by this API.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/collectionInfo'
            text/html:
              schema:
                type: string
        default:
          description: An error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
...
components:
  schemas:
    exception:
      required:
        - code
      type: object
      properties:
        code:
          type: string
        description:
          type: string
...
    link:
      required:
        - href
      type: object
      properties:
        href:
          type: string
        rel:
          type: string
          example: prev
        type:
          type: string
          example: application/geo+json
        hreflang:
          type: string
          example: en
...
    collectionInfo:
      required:
        - links
        - name
      type: object
      properties:
        name:
          type: string
          description: 'identifier of the collection used, for example, in URIs'
          example: buildings
        title:
          type: string
          description: human readable title of the collection
          example: Buildings
        description:
          type: string
          description: a description of the data in the collection
          example: Buildings in the city of Bonn.
        links:
          type: array
          example:
            - href: 'http://data.example.org/collections/buildings/items'
              rel: item
              type: application/geo+json
              title: Buildings
            - href: 'http://example.org/concepts/building.html'
              rel: describedBy
              type: text/html
              title: Coverage for buildings
          items:
            $ref: '#/components/schemas/link'
        extent:
          $ref: '#/components/schemas/extent'
        crs:
          type: array
          description: 'The coordinate reference systems in which geometries may be retrieved. Coordinate reference systems are identified by a URI. The first coordinate reference system is the coordinate reference system that is used by default. This is always "http://www.opengis.net/def/crs/OGC/1.3/CRS84", i.e. WGS84 longitude/latitude.'
          items:
            type: string
          default:
            - 'http://www.opengis.net/def/crs/OGC/1.3/CRS84'
    extent:
      type: object
      properties:
        crs:
          type: string
          description: 'Coordinate reference system of the coordinates in the spatial extent (property `spatial`). In the Core, only WGS84 longitude/latitude is supported. Extensions may support additional coordinate reference systems.'
          default: 'http://www.opengis.net/def/crs/OGC/1.3/CRS84'
          enum:
            - 'http://www.opengis.net/def/crs/OGC/1.3/CRS84'
        spatial:
          maxItems: 6
          minItems: 4
          type: array
          description: 'West, north, east, south edges of the spatial extent. The minimum and maximum values apply to the coordinate reference system WGS84 longitude/latitude that is supported in the Core. If, for example, a projected coordinate reference system is used, the minimum and maximum values need to be adjusted.'
          example:
            - -180
            - -90
            - 180
            - 90
          items:
            type: number
        trs:
          type: string
          description: 'Temporal reference system of the coordinates in the temporal extent (property `temporal`). In the Core, only the Gregorian calendar is supported. Extensions may support additional temporal reference systems.'
          default: 'http://www.opengis.net/def/uom/ISO-8601/0/Gregorian'
          enum:
            - 'http://www.opengis.net/def/uom/ISO-8601/0/Gregorian'
        temporal:
          maxItems: 2
          minItems: 2
          type: array
          description: Begin and end times of the temporal extent.
          example:
            - 2011-11-11T12:22:11.000Z
            - 2012-11-24T12:32:43.000Z
          items:
            type: string
            format: dateTime
...